+++ src/components/code-snippet/code-snippet.js
@@ -7,6 +7,7 @@ import CodeSnippetTitle from '../code-snippet-title';
 import Edit from '../edit';
 import { highlightThemeCss } from '../highlight/theme-css.js';
 import classnames from 'classnames';
+import onCopy from './on-copy';
 
 class CodeSnippet extends React.PureComponent {
   editButtons = () => {
@@ -42,9 +43,10 @@ class CodeSnippet extends React.PureComponent {
     const filename = this.props.filename;
     return filename && <CodeSnippetTitle filename={filename} />;
   };
+
   render() {
     const { code, maxHeight, highlighter } = this.props;
-    const highlightedCode = highlighter()(code);
+    const highlightedCode = highlighter(code);
     // only load the component if we have `code` and `highlighter`
     if (!highlighter && !code) return;
     // wrap the component in appcontext so we can get the user's token
@@ -59,9 +61,7 @@ class CodeSnippet extends React.PureComponent {
             {...this.props}
             highlightThemeCss={highlightThemeCss}
             highlightedCode={highlightedCode}
-            onCopy={() => {
-              analytics.track('Copied example with clipboard');
-            }}
+            onCopy={onCopy}
           />
         </div>
       </div>
@@ -72,7 +72,7 @@ class CodeSnippet extends React.PureComponent {
 CodeSnippet.propTypes = {
   /** The raw code. */
   code: PropTypes.string.isRequired,
-  /** The dr-ui/highlight function, example: "highlighter={() => highlightJson}". You must also import the function in your frontmatter. */
+  /** The dr-ui/highlight function, example: "highlighter={highlightJson}". You must also import the function in your frontmatter. */
   highlighter: PropTypes.func.isRequired,
   /** Name of the file to add context to the code block. */
   filename: PropTypes.string,