+++ src/components/numbered-code-snippet/numbered-code-snippet.js
@@ -4,6 +4,7 @@ import debounce from 'debounce';
 import classnames from 'classnames';
 import CopyButton from '@mapbox/mr-ui/copy-button';
 import { HideLines, ShowLines } from './show-hide-lines';
+import onCopy from '../code-snippet/on-copy';
 
 function getWindow() {
   if (typeof window === undefined) {
@@ -136,6 +137,10 @@ export default class NumberedCodeSnippet extends React.PureComponent {
     this.firstLiveElement = element;
   };
 
+  toggleLines = () => {
+    this.setState({ expanded: !this.state.expanded });
+  };
+
   render() {
     const { props } = this;
 
@@ -249,7 +254,6 @@ export default class NumberedCodeSnippet extends React.PureComponent {
     const highlightElements = [];
     const copyElements = [];
     let previousCount = 0;
-    let liveChunkCount = -1; // Incremented to give CopyButtons an identifier
     allChunks.forEach((codeChunk, i) => {
       const chunkId = `chunk-${i}`;
       const lineEls = codeChunk.highlightedLines.map((line, i) => {
@@ -340,19 +344,9 @@ export default class NumberedCodeSnippet extends React.PureComponent {
         );
       } else if (!codeChunk.live && lineEls.length) {
         const expandCollapseButtons = this.state.expanded ? (
-          <HideLines
-            onClick={() => {
-              this.setState({ expanded: !this.state.expanded });
-            }}
-          >
-            {lineEls}
-          </HideLines>
+          <HideLines onClick={this.toggleLines}>{lineEls}</HideLines>
         ) : (
-          <ShowLines
-            onClick={() => {
-              this.setState({ expanded: !this.state.expanded });
-            }}
-          />
+          <ShowLines onClick={this.toggleLines} />
         );
         codeElements.push(
           <div
@@ -383,21 +377,16 @@ export default class NumberedCodeSnippet extends React.PureComponent {
           </div>
         );
 
-        const chunkIndex = ++liveChunkCount;
-        const onCopyChunk = () => this.props.onCopy(chunkIndex);
-
-        if (props.onCopy) {
-          copyElements.push(
-            <div
-              key={i}
-              data-chunk-copy={chunkId}
-              className="absolute z3 right mr3 color-white"
-              style={{ opacity: 0, transition: 'opacity 300ms linear' }}
-            >
-              <CopyButton text={codeChunk.raw} onCopy={onCopyChunk} />
-            </div>
-          );
-        }
+        copyElements.push(
+          <div
+            key={i}
+            data-chunk-copy={chunkId}
+            className="absolute z3 right mr3 color-white"
+            style={{ opacity: 0, transition: 'opacity 300ms linear' }}
+          >
+            <CopyButton text={codeChunk.raw} onCopy={onCopy} />
+          </div>
+        );
       }
       previousCount = previousCount + lineEls.length;
     });
@@ -406,10 +395,10 @@ export default class NumberedCodeSnippet extends React.PureComponent {
     const codeClasses = 'px0 hljs';
 
     let copyAllButton = null;
-    if (props.copyRanges === undefined && props.onCopy) {
+    if (props.copyRanges === undefined) {
       copyAllButton = (
         <div className="absolute z2 top right mr6 mt6 color-white">
-          <CopyButton text={props.code} onCopy={props.onCopy} />
+          <CopyButton text={props.code} onCopy={onCopy} />
         </div>
       );
     }
@@ -463,11 +452,6 @@ NumberedCodeSnippet.propTypes = {
    * A maximum height for the snippet. If the code exceeds this height, the snippet will scroll internally.
    */
   maxHeight: PropTypes.number,
-  /**
-   * A callback that is invoked when the snippet (or a chunk of the snippet) is copied. If `copyRanges`
-   * are provided, the callback is passed the index (0-based) of the chunk that was copied.
-   */
-  onCopy: PropTypes.func,
   /**
    * CSS that styles the highlighted code.
    */